Moving unittests that used tools to functional (#8250)
* remove visual_studio marks

* mark tests that require pkg-config

* mark test as gcc

* move msbuild tests

* move win tests to functional

* move scm, git, svn to functional

* move scm tests

* move tests

* move compiler detection tests

* move visual_studio tests to functional

* mv pkgconfig tests

* move file test
