+++ conans/test/functional/client/tools/win/vcvars_clangcl_test.py
