+++ conans/test/functional/client/tools/win/vcvars_store_test.py
