+++ conans/test/functional/client/tools/win/vcvars_test.py
