+++ conans/test/functional/util/pkg_config_test.py
@@ -39,6 +39,7 @@ class PkgConfigTest(unittest.TestCase):
         with self.assertRaises(ConanException):
             pc.libs()
 
+    @pytest.mark.tool_pkg_config
     def test_pc(self):
         if platform.system() == "Windows":
             return
@@ -63,6 +64,7 @@ class PkgConfigTest(unittest.TestCase):
             self.assertEqual(frozenset(pkg_config.version), frozenset(['6.6.6']))
         os.unlink(filename)
 
+    @pytest.mark.tool_pkg_config
     def test_define_prefix(self):
         if platform.system() == "Windows":
             return
@@ -89,6 +91,7 @@ class PkgConfigTest(unittest.TestCase):
             self.assertEqual(frozenset(pkg_config.version), frozenset(['6.6.6']))
         os.unlink(filename)
 
+    @pytest.mark.tool_pkg_config
     def test_rpaths_libs(self):
         if platform.system() == "Windows":
             return
