+++ conans/test/unittests/util/msvs_toolset_test.py
@@ -1,19 +1,14 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-import platform
 import unittest
 
-import pytest
-from nose.plugins.attrib import attr
 from parameterized import parameterized
 
 from conans.client import tools
 from conans.test.utils.mocks import MockSettings
 
 
-@pytest.mark.tool_visual_studio
-@unittest.skipUnless(platform.system() == "Windows", "Requires Windows")
 class MSVCToolsetTest(unittest.TestCase):
 
     @parameterized.expand([("16", "v142"),
